Revert "LLVMLinux: Apply all necessary patches"
This reverts commit 747989a0858f7499e36dc646b83835beb43c9495.

+++ arch/arm64/include/asm/arch_timer.h
@@ -31,12 +31,10 @@ static inline void arch_timer_reg_write_cp15(int access, int reg, u32 val)
 	if (access == ARCH_TIMER_PHYS_ACCESS) {
 		switch (reg) {
 		case ARCH_TIMER_REG_CTRL:
-			asm volatile("msr cntp_ctl_el0,  %0"
-				: : "r" ((u64)val));
+			asm volatile("msr cntp_ctl_el0,  %0" : : "r" (val));
 			break;
 		case ARCH_TIMER_REG_TVAL:
-                       asm volatile("msr cntp_tval_el0, %0"
-                               : : "r" ((u64)val));
+			asm volatile("msr cntp_tval_el0, %0" : : "r" (val));
 			break;
 		default:
 			BUILD_BUG();
@@ -44,14 +42,10 @@ static inline void arch_timer_reg_write_cp15(int access, int reg, u32 val)
 	} else if (access == ARCH_TIMER_VIRT_ACCESS) {
 		switch (reg) {
 		case ARCH_TIMER_REG_CTRL:
-                       asm volatile("msr cntv_ctl_el0,  %0"
-                               : : "r" ((u64)val));
-
+			asm volatile("msr cntv_ctl_el0,  %0" : : "r" (val));
 			break;
 		case ARCH_TIMER_REG_TVAL:
-                       asm volatile("msr cntv_tval_el0, %0"
-                               : : "r" ((u64)val));
-
+			asm volatile("msr cntv_tval_el0, %0" : : "r" (val));
 			break;
 		default:
 			BUILD_BUG();
@@ -65,7 +59,7 @@ static inline void arch_timer_reg_write_cp15(int access, int reg, u32 val)
 
 static inline u32 arch_timer_reg_read_cp15(int access, int reg)
 {
-	u64 val;
+	u32 val;
 
 	if (access == ARCH_TIMER_PHYS_ACCESS) {
 		switch (reg) {
@@ -93,26 +87,26 @@ static inline u32 arch_timer_reg_read_cp15(int access, int reg)
 		BUILD_BUG();
 	}
 
-	return (u32)val;
+	return val;
 }
 
 static inline u32 arch_timer_get_cntfrq(void)
 {
-	u64 val;
+	u32 val;
 	asm volatile("mrs %0,   cntfrq_el0" : "=r" (val));
-	return (u32)val;
+	return val;
 }
 
 static inline u32 arch_timer_get_cntkctl(void)
 {
-	u64 cntkctl;
+	u32 cntkctl;
 	asm volatile("mrs	%0, cntkctl_el1" : "=r" (cntkctl));
-	return (u32)cntkctl;
+	return cntkctl;
 }
 
 static inline void arch_timer_set_cntkctl(u32 cntkctl)
 {
-	asm volatile("msr	cntkctl_el1, %0" : : "r" ((u64)cntkctl));
+	asm volatile("msr	cntkctl_el1, %0" : : "r" (cntkctl));
 }
 
 static inline void arch_timer_evtstrm_enable(int divider)
