+++ arch/arm64/kernel/debug-monitors.c
@@ -52,15 +52,15 @@ static void mdscr_write(u32 mdscr)
 {
 	unsigned long flags;
 	local_dbg_save(flags);
-	asm volatile("msr mdscr_el1, %0" : : "r" ((u64)mdscr));
+	asm volatile("msr mdscr_el1, %0" :: "r" (mdscr));
 	local_dbg_restore(flags);
 }
 
 static u32 mdscr_read(void)
 {
-	u64 mdscr;
+	u32 mdscr;
 	asm volatile("mrs %0, mdscr_el1" : "=r" (mdscr));
-	return (u32)mdscr;
+	return mdscr;
 }
 
 /*
@@ -136,7 +136,7 @@ void disable_debug_monitors(enum debug_el el)
  */
 static void clear_os_lock(void *unused)
 {
-	asm volatile("msr oslar_el1, %0" : : "r" ((u64)0));
+	asm volatile("msr oslar_el1, %0" : : "r" (0));
 }
 
 static int __cpuinit os_lock_notify(struct notifier_block *self,
